create project

+++ src/main/java/com/basharina/taskmanagementsystem/repository/CommentRepository.java
@@ -3,5 +3,5 @@ package com.basharina.taskmanagementsystem.repository;
 import com.basharina.taskmanagementsystem.model.entity.CommentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentRepository extends JpaRepository<CommentEntity, Integer> {
+public interface CommentRepository extends JpaRepository<CommentEntity, Long> {
 }
